[fli-iam#2059] VIP API client + proxy APIs in back
