+++ shanoir-ng-datasets/src/main/resources/application.yml
@@ -136,7 +136,7 @@ dcm4chee-arc:
 
 # VIP
 vip:
-  uri: ${VIP_URL_SCHEME}://${VIP_URL_HOST}/rest/executions/
+  uri: ${VIP_URL_SCHEME}://${VIP_URL_HOST}/rest/
   upload-folder: /tmp/vip_uploads
   sleep-time: 20000
   file-formats: .nii.gz, .nii
